chore: move http to typescript, small refactor
