+++ packages/browser/lib/core/http.ts
@@ -1,0 +1,46 @@
+import type { HttpConfig } from "../config";
+
+import { errors } from "../utils";
+
+export default function Http(config: HttpConfig, teamId: string, appId: string, context: string) {
+  if(window == null) {
+    throw new errors.InitializationError(
+      "\`window\` object not found. You cannot run this SDK outside of a browser environment."
+    );
+  };
+
+  if (!("fetch" in window)) {
+    throw new errors.InitializationError(
+      "Your browser is outdated and does not support window.fetch(). Please upgrade it."
+    );
+  };
+
+  const getCageKey = async () => {
+    try {
+      const keyEndpoint = new URL(`${teamId}/apps/${appId}?context=${context}`, config.keysUrl);
+
+      const response = await fetch(keyEndpoint, {
+        method: "GET",
+      });
+
+      if (!response.ok) {
+        throw new Error(`HTTP error! status: ${response.status}`);
+      }
+
+      const body = await response.json();
+      const headers = response.headers;
+
+      return {
+        ...body,
+        isDebugMode: headers.get("X-Evervault-Inputs-Debug-Mode") === "true",
+      };
+    } catch (err) {
+      throw new errors.CageKeyError(
+        "An error occurred while retrieving the cage's key",
+        { cause: err }
+      );
+    }
+  };
+
+  return { getCageKey };
+}
